[Improve] moduleにTwitterAPI部分を切り出し
